add integration tests

+++ Dynamiq.API/Dynamiq.API/Controllers/UserController.cs
@@ -1,8 +1,11 @@
 ﻿using Dynamiq.Application.Commands.Users.Commands;
 using Dynamiq.Application.Queries.Users.Queries;
+using Dynamiq.Domain.Enums;
+using Dynamiq.Infrastructure.Services.Auth;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace Dynamiq.API.Controllers
 {
@@ -21,18 +24,27 @@ namespace Dynamiq.API.Controllers
 
         [Authorize]
         [HttpGet("{id}")]
-        public async Task<IActionResult> GetById([FromRoute] Guid id)
+        public async Task<IActionResult> GetById([FromQuery] Guid id)
         {
             var user = await _mediator.Send(new GetUserByIdQuery(id));
+
+            if(user == null) 
+                return NotFound();
 
             _logger.LogInformation("Retrieved user with id: {Id}", id);
 
             return Ok(user);
         }
 
-        [HttpPut("changePassword")]
+        [Authorize]
+        [HttpPut("change-password")]
         public async Task<IActionResult> Put([FromBody] ChangeUserPasswordCommand command)
         {
+            var userEmail = User.FindFirst(ClaimTypes.Email)?.Value;
+
+            if (userEmail != command.Email)
+                return Forbid();
+
             await _mediator.Send(command);
 
             _logger.LogInformation("Changed password for user with email: {Email}", command.Email);
@@ -40,14 +52,21 @@ namespace Dynamiq.API.Controllers
             return Ok("You successfully changed your password");
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
-        public async Task<IActionResult> Delete([FromRoute] Guid id)
+        public async Task<IActionResult> Delete(Guid id)
         {
+            var usersId = User.FindFirst(JwtClaims.UserId)?.Value;
+            var usersRole = User.FindFirst(ClaimTypes.Role)?.Value;
+
+            if (usersId != id.ToString() || usersRole == RoleEnum.Admin.ToString())
+                return Forbid();
+
             await _mediator.Send(new DeleteUserCommand(id));
 
             _logger.LogInformation("Deleted user with id: {Id}", id);
 
-            return Ok("DefaultUser was removed");
+            return Ok("user was removed");
         }
     }
 }
